Bug huntin'

+++ soundenc/quantizer.cc
@@ -31,7 +31,7 @@
 
 #define FRAME_SIZE 512
 
-unsigned int quant_sum = 0;
+float quant_sum = 0;
 
 void quantize(float thres, float max[], int b, int s, samples_t *band)
 {
